Add files via upload
The purpose of this work is to understand how digital signal processing works in real
life. In our case we will take an analog input(a sound) and convert it with the help of
Fast Fourier transform, which converts a signal into individual spectral components
and thereby provides frequency information about the signal.